Migrate models, tests and consumers
