+++ src/arena/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from arena.models import Room, Player
+
+# Register your models here.
+admin.site.register(Room)
+admin.site.register(Player)
