+++ src/arena/routing.py
@@ -4,5 +4,5 @@ from . import consumers
 
 
 websocket_urlpatterns = [
-
+    re_path(r"ws/arena/(?P<room_id>\w+)", consumers.ArenaConsumer.as_asgi()),
 ]
