+++ src/arena/tests/test_consumers/test_consumer.py
@@ -8,7 +8,7 @@ from rest_framework_simplejwt.tokens import AccessToken
 
 from app.asgi import application
 
-from lobby.models import Room, Player
+from arena.models import Room, Player
 
 
 TEST_CHANNEL_LAYERS = {
